refactor(verify): `zk_resolve_context` moved
