+++ crates/verify/src/verify.rs
@@ -4,7 +4,7 @@ use crate::{
     etherscan::EtherscanVerificationProvider,
     provider::{VerificationContext, VerificationProvider, VerificationProviderType},
     utils::is_host_only,
-    zk_provider::{CompilerVerificationContext, ZkVerificationContext},
+    zk_provider::CompilerVerificationContext,
     RetryArgs,
 };
 use alloy_primitives::Address;
@@ -23,6 +23,8 @@ use reqwest::Url;
 use revm_primitives::HashSet;
 use semver::BuildMetadata;
 use std::path::PathBuf;
+
+mod zksync;
 
 /// Verification provider arguments
 #[derive(Clone, Debug, Parser)]
@@ -208,11 +210,7 @@ impl VerifyArgs {
             None => config.chain.unwrap_or_default(),
         };
 
-        let context = if self.zksync {
-            CompilerVerificationContext::ZkSolc(self.zk_resolve_context().await?)
-        } else {
-            CompilerVerificationContext::Solc(self.resolve_context().await?)
-        };
+        let context = self.resolve_context().await?;
 
         // Set Etherscan options.
         self.etherscan.chain = Some(chain);
@@ -268,7 +266,19 @@ impl VerifyArgs {
 
     /// Resolves [VerificationContext] object either from entered contract name or by trying to
     /// match bytecode located at given address.
-    pub async fn resolve_context(&self) -> Result<VerificationContext> {
+    pub async fn resolve_context(&self) -> Result<CompilerVerificationContext> {
+        if self.zksync {
+            self.zk_resolve_context().await.map(CompilerVerificationContext::ZkSolc)
+        } else {
+            self.resolve_solc_context().await.map(CompilerVerificationContext::Solc)
+        }
+    }
+
+    /// Resolves [VerificationContext] object either from entered contract name or by trying to
+    /// match bytecode located at given address.
+    ///
+    /// Will assume configured compiler is solc or equivalent
+    async fn resolve_solc_context(&self) -> Result<VerificationContext> {
         let mut config = self.load_config_emit_warnings();
         config.libraries.extend(self.libraries.clone());
 
@@ -400,81 +410,6 @@ impl VerifyArgs {
                 artifact_id.version.clone(),
                 config,
                 settings.clone(),
-            )
-        }
-    }
-
-    pub async fn zk_resolve_context(&self) -> Result<ZkVerificationContext> {
-        let mut config = self.load_config_emit_warnings();
-        config.libraries.extend(self.libraries.clone());
-
-        let project = foundry_config::zksync::config_create_project(&config, config.cache, false)?;
-
-        if let Some(ref contract) = self.contract {
-            let contract_path = if let Some(ref path) = contract.path {
-                project.root().join(PathBuf::from(path))
-            } else {
-                project.find_contract_path(&contract.name)?
-            };
-
-            let version = if let Some(ref version) = self.compiler_version {
-                version.trim_start_matches('v').parse()?
-            } else if let Some(ref solc) = config.solc {
-                match solc {
-                    SolcReq::Version(version) => version.to_owned(),
-                    SolcReq::Local(solc) => Solc::new(solc)?.version,
-                }
-            } else if let Some(entry) = project
-                .read_cache_file()
-                .ok()
-                .and_then(|mut cache| cache.files.remove(&contract_path))
-            {
-                let unique_versions = entry
-                    .artifacts
-                    .get(&contract.name)
-                    .map(|artifacts| artifacts.keys().collect::<HashSet<_>>())
-                    .unwrap_or_default();
-
-                if unique_versions.is_empty() {
-                    eyre::bail!("No matching artifact found for {}", contract.name);
-                } else if unique_versions.len() > 1 {
-                    warn!(
-                        "Ambiguous compiler versions found in cache: {}",
-                        unique_versions.iter().join(", ")
-                    );
-                    eyre::bail!("Compiler version has to be set in `foundry.toml`. If the project was not deployed with foundry, specify the version through `--compiler-version` flag.")
-                }
-
-                unique_versions.into_iter().next().unwrap().to_owned()
-            } else {
-                eyre::bail!("If cache is disabled, compiler version must be either provided with `--compiler-version` option or set in foundry.toml")
-            };
-
-            ZkVerificationContext::new(contract_path, contract.name.clone(), version, config)
-        } else {
-            if config.get_rpc_url().is_none() {
-                eyre::bail!("You have to provide a contract name or a valid RPC URL")
-            }
-            let provider = utils::get_provider(&config)?;
-            let code = provider.get_code_at(self.address).await?;
-
-            let output = ProjectCompiler::new().zksync_compile(&project)?;
-            let contracts = ContractsByArtifact::new(
-                output.artifact_ids().map(|(id, artifact)| (id, artifact.clone().into())),
-            );
-
-            let Some((artifact_id, _)) = contracts.find_by_deployed_code_exact(&code) else {
-                eyre::bail!(format!(
-                    "Bytecode at {} does not match any local contracts",
-                    self.address
-                ))
-            };
-
-            ZkVerificationContext::new(
-                artifact_id.source.clone(),
-                artifact_id.name.split('.').next().unwrap().to_owned(),
-                artifact_id.version.clone(),
-                config,
             )
         }
     }
